Simplify code to make it easier to change
As hacking the vim-wasm packages is required, I'm not going to be able to
benefit from the tests as easy as if I were to make the changes "properly".

Perhaps after this spike, if it works, I can do this "properly" and open a PR
but for now I want to know if my approach will work before I put in the extra
effort.