Detect if you've won or lost
For some reason, once you call `:vs` via `vim.cmdline`, it doesn't reflect in
the vim UI until another key is detected but the `:redraw` command fixes this.

Also, to prevent multiple submissions, I'm quiting vim at the end.
This means I'll need to have a way of starting vim again (rerun effect of
re-render?).

Right now if the user submits a wrong answer, the diff shows them where they
went wrong.
Another option would be to allow the user to continue adding keystrokes and
submit again and I wouldn't need to quit but this has the downsides of not
showing them the diff and it wouldn't result in the best score (golf).

This may be useful for the class room situation, although not if the challenge
is "transform this text in under X keystrokes".