Add syntax highlighting to the target window
This uses the same colorscheme as the vim instance, although the colors are
slightly different, it's close enough for now.

Also, importing the theme from react-syntax-highlighter seemed to throw a syntax
error, not quite sure why but copying and pasting it will make it easier to
tweak later on to make it match vim better.
I already changed the fontSize to match the rest and it worked well.

Closes #3